core: clean up methods and tests
